Move rest of the ui components to the screenshot module.
