+++ app/src/main/java/com/appunite/loudius/ui/authenticating/AuthenticatingScreen.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.appunite.loudius.R
 import com.appunite.loudius.screenshots.components.LoudiusFullScreenError
+import com.appunite.loudius.screenshots.components.LoudiusLoadingIndicator
 import com.appunite.loudius.screenshots.theme.LoudiusTheme
-import com.appunite.loudius.ui.components.LoudiusLoadingIndicator
 import com.appunite.loudius.screenshots.R as screenshotsR
 
 @Composable
