+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,5 @@
 <resources>
     <string name="app_name">Loudius</string>
-    <string name="back_button">Back button</string>
     <string name="pull_requests_screen_pull_request_content_description">Pull request</string>
     <string name="details_screen_user_image_description">User image</string>
     <string name="details_notify">Notify</string>
@@ -12,7 +11,6 @@
     <string name="reviewers_snackbar_success">Awesome! Your collaborator have been pinged for some serious code review action! 🎉</string>
     <string name="reviewers_snackbar_failure">Uh-oh, it seems that Loudius has taken a vacation. Don\'t worry, we\'re sending a postcard to bring it back ASAP!</string>
     <string name="user_unauthorized_message">Unauthorized collaborator detected! Please login again.</string>
-    <string name="you_dont_have_any_pull_request">Sorry! Your list of pull requests is empty.\nGet back to work! 🧑‍💻</string>
     <string name="you_dont_have_any_reviewers">Sorry! Your list of reviewers is empty.\n Go to pull request and mark your colleagues as the reviewers! 🤞</string>
 
     <!-- LoginScreen -->
