+++ screenshots/build.gradle
@@ -36,6 +36,9 @@ dependencies {
     implementation 'androidx.compose.ui:ui-tooling-preview'
     debugImplementation 'androidx.compose.ui:ui-tooling'
 
+    //Lottie - Compose
+    implementation("com.airbnb.android:lottie-compose:5.2.0")
+
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     testImplementation 'junit:junit:4.13.2'
