+++ screenshots/src/main/java/com/appunite/loudius/screenshots/common/CountingIdlingResource.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.appunite.loudius.common
+package com.appunite.loudius.screenshots.common
 
 import java.util.concurrent.atomic.AtomicInteger
