+++ screenshots/src/main/java/com/appunite/loudius/screenshots/components/LoudiusLoadingIndicator.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.appunite.loudius.ui.components
+package com.appunite.loudius.screenshots.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -30,7 +30,7 @@ import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.LottieConstants
 import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
-import com.appunite.loudius.R
+import com.appunite.loudius.screenshots.R
 import com.appunite.loudius.screenshots.theme.LoudiusTheme
 
 @Composable
