+++ screenshots/src/main/res/drawable/arrow_back.xml
