+++ screenshots/src/main/res/values/strings.xml
@@ -1,8 +1,10 @@
 <resources>
+    <string name="back_button">Back button</string>
     <string name="error_dialog_title">Error</string>
     <string name="error_dialog_text">Something went wrong…</string>
-    <string name="ok">OK</string>
     <string name="error_image_desc">error image</string>
-    <string name="try_again">Try again</string>
     <string name="error_login_text">Something went wrong…\nYou need to log in again.</string>
+    <string name="ok">OK</string>
+    <string name="try_again">Try again</string>
+    <string name="you_dont_have_any_pull_request">Sorry! Your list of pull requests is empty.\nGet back to work! 🧑‍💻</string>
 </resources>
